Add skip_werror macro
This patch add skip_werror macro for deleting '-Werror' in the CFLAGS,
when build fail cases by Warning.

For example, when we build it on riscv64, libfdt_wrap.c is generated by
swig, but it cases some build error like below:

/builddir/build/BUILD/dtc-1.6.1/pylibfdt/libfdt_wrap.c:9491:1: note: in expansion of macro ‘SWIG_init’
 9491 | SWIG_init(void) {
      | ^~~~~~~~~
cc1: all warnings being treated as errors

For building this package in testing purpose, we can use
'--with skip_werror' to build the package.

Signed-off-by: Wei Fu <wefu@redhat.com>